Fix RegistryClass assert

+++ Code/Libraries/Source/WWVegas/WWLib/registry.cpp
@@ -47,12 +47,12 @@ RegistryClass::RegistryClass( const char * sub_key ) :
 {
 	DWORD disposition;
 	HKEY	key;
-	assert( sizeof(HKEY) == sizeof(int) );
+	assert( sizeof(HKEY) == sizeof(intptr_t) );
 	if (::RegCreateKeyEx( HKEY_LOCAL_MACHINE, sub_key,
 			0, NULL, 0, KEY_ALL_ACCESS, NULL,
 			&key, &disposition ) == ERROR_SUCCESS) {
 		IsValid = true;
-		Key = (int)key;
+		Key = (intptr_t)key;
 	}
 }
